pyproject.toml
[tool.poetry]
name = "discord-management-bot2"
version = "0.1.0"
description = ""
authors = ["jukey17 <keykun.pv@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_management_bot2"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
python-dotenv = "^0.21.0"
pytz = "^2022.4"
pydantic = "^1.10.2"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
